+++ app/src/main/java/io/skyvoli/goodbooks/storage/database/dao/BookDao.java
@@ -9,15 +9,11 @@ import androidx.room.Update;
 import java.util.List;
 
 import io.skyvoli.goodbooks.storage.database.dto.Book;
-import io.skyvoli.goodbooks.ui.fragments.series.Series;
 
 @Dao
 public interface BookDao {
     @Query("SELECT * FROM books ORDER BY title, part")
     List<Book> getAll();
-
-    @Query("SELECT title, COUNT(*) as countedBooks FROM books GROUP BY title")
-    List<Series> getSeries();
 
     @Insert
     void insert(Book... books);
